Move room and client to separate files
