+++ src/server/client.cpp
@@ -1,0 +1,9 @@
+#include "client.hpp"
+
+#include <arpa/inet.h>
+#include <string>
+
+Client::~Client() {
+    delete this->address;
+}
+
